Fix lollipop bug
Fix a necessitas bug for Android Lollipop. The framework is unable to
create a assets file engine, which leads to a fatal error thereafter.
That happens because the assets file engine creator method expects an
environment variable to create such engine but the method to retrieve
that information stopped working in Android Lollipop.

The solution was to get rid of that variable dependency and create a
assets engine without this information.

Signed-off-by: Mailson Menezes <mailson.menezes@openbossa.org>